# 60x Speedup on IUV2FBC Function. Greatly speeds up converting a 2D person image to a 3D model.