… DensePoseData/demo_data_2/INDS.png
… DensePoseData/demo_data_2/IUV.png
… DensePoseData/demo_data_2/im.jpg
… notebooks/DensePose-Fast-IUV-2-XYZ.ipynb