Hello,

This PR gives a 60x speed up to ```IUV2FBC``` (_in DensePose/detectron/utils/densepose_methods.py_). ```IUV2FBC``` function converts an IUV point to an XYZ point on a SMPL model. This will convenience users who would like to make a 3D person out of a 2D person image (e.g. in virtual reality)

## Timings

### Run IUV2FBC on 125 IUV points in 'DensePoseData/demo_data/demo_dp_single_ann.pkl' ...

**Original:**
```
8629403 function calls in 5.227 seconds
Ordered by: cumulative time
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    5.227    5.227 <string>:1(<module>)
        1    0.001    0.001    5.227    5.227 <ipython-input-40-02014517a6cb>:32(f)
      125    0.088    0.001    5.219    0.042 densepose_methods.py:270(IUV2FBC)
    35578    0.430    0.000    5.064    0.000 densepose_methods.py:62(barycentric_coordinates_exists)
   108588    1.486    0.000    4.536    0.000 numeric.py:1591(cross)
```
**Sped up:**
```
39129 function calls in 0.081 seconds
Ordered by: cumulative time
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.081    0.081 <string>:1(<module>)
        1    0.001    0.001    0.081    0.081 <ipython-input-39-6b6ed951a68c>:32(f)
      125    0.014    0.000    0.079    0.001 densepose_methods.py:174(IUV2FBC_fast)
      125    0.035    0.000    0.059    0.000 densepose_methods.py:135(barycentric_coordinates_fast)
      375    0.008    0.000    0.020    0.000 numeric.py:1591(cross)
```

### Run IUV2FBC on 80000 IUV points ...

![summary](https://user-images.githubusercontent.com/1690072/44390932-1951c480-a561-11e8-98f1-1242f7790c94.png)

**Original:** Estimated ~53 minutes (80000 points / 125 points * 5 secs ~= 53 mins)

**Sped up:** 1 minute (timing obtained from a profiled run)

## PR Contribution

notebooks/DensePose-Fast-IUV-2-XYZ.ipynb :  Is a demo. Also addresses issue #87 .

detectron/utils/densepose_methods.py : Contains sped-up codes. Backward-compatible since old functions not modified. New functions: ```IUV2FBC_fast``` and ```barycentric_coordinates_fast```.

DensePoseData/demo_data_2/* : Data given for the convenience of running DensePose-Fast-IUV-2-XYZ.ipynb .