@davidleejy 2018-09-19
@nvrv @ralpguler 
Loved using Densepose and would be grateful for the opportunity to contribute to this repository. The following sheds more light on how the speed-up came about:

- Main source of speed-up is **batching numpy cross product computations** (https://stackoverflow.com/questions/1988091/poor-numpy-cross-performance) executed frequented in barycentric_coordinates_exists() and barycentric_coordinates(). Instead of calling numpy.cross() on 2 vectors at a time, I called it on a bunch of vectors in a single call.

- Secondary source of speed-up is **removing duplicate computations** in barycentric_coordinates_exists() and barycentric_coordinates().

- The result is a single function barycentric_coordinates_fast().

- New code is functionally identical to original code in that it produces the same results.

- **End-users** call a new function IUV2FBC_fast() that **takes in the same parameters as the original function** IUV2FBC().  
_IUV2FBC_fast( self, I_point , U_point, V_point)_ 
_IUV2FBC( self, I_point , U_point, V_point)_

- I admit the **demo notebook** I gave is kind of messy in hindsight. I could have removed a lot of cells to help better convey this improvement. I could clean it up if it'd help the reviewers & maintainers. :)

Feel free to comment with questions and/or suggestions :)